Removed common module and create ConnectionInterceptor
