+++ datasource-remote/build.gradle.kts
@@ -48,7 +48,6 @@ android {
 
 dependencies {
     api(project(":data"))
-    implementation (project(":common"))
     implementation(libs.androidx.core.ktx)
     implementation(platform(libs.koin.bom))
     implementation(libs.koin.android)
