+++ datasource-remote/src/main/java/cmm/apps/esmorga/datasource_remote/api/ExceptionHandler.kt
@@ -1,7 +1,6 @@
 package cmm.apps.esmorga.datasource_remote.api
 
 import android.content.Context
-import cmm.apps.esmorga.common.util.ConnectivityUtils
 import cmm.apps.esmorga.domain.result.ErrorCodes
 import cmm.apps.esmorga.domain.result.EsmorgaException
 import cmm.apps.esmorga.domain.result.Source
@@ -32,7 +31,6 @@ object ExceptionHandler {
 
             is ConnectException,
             is UnknownHostException -> {
-                ConnectivityUtils.reportNoConnectivityIfNeeded(context)
                 EsmorgaException(
                     message = "No connection error: ${e.message.orEmpty()}",
                     source = Source.REMOTE,
