+++ settings.gradle.kts
@@ -27,4 +27,3 @@ include(":datasource-remote")
 include(":view")
 include(":datasource-local")
 include(":designsystem")
-include(":common")
