+++ view/build.gradle.kts
@@ -46,7 +46,6 @@ dependencies {
 
     api(project(":domain"))
     implementation(project(":designsystem"))
-    implementation (project(":common"))
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
     implementation(libs.androidx.activity.compose)
