+++ view/src/test/java/cmm/apps/esmorga/view/screenshot/profile/ProfileViewScreenshotTest.kt
@@ -33,7 +33,6 @@ class ProfileViewScreenshotTest : BaseScreenshotTest() {
                         user = user
                     ),
                     shownLogOutDialog = {},
-                    onChangePassword = {},
                     onNavigateLogin = {}
                 )
             }
